requirements.txt
requests
dnslib
aiosmtpd
pycryptodome